+++ sdk/metric/doc.go
@@ -28,9 +28,7 @@ and asynchronous instruments.  There are two constructors per instrument for
 the two kinds of number (int64, float64).
 
 Synchronous instruments are managed by a sync.Map containing a *record
-with the current state for each synchronous instrument.  A bound
-instrument encapsulates a direct pointer to the record, allowing
-bound metric events to bypass a sync.Map lookup.  A lock-free
+with the current state for each synchronous instrument.  A lock-free
 algorithm is used to protect against races when adding and removing
 items from the sync.Map.
 
@@ -45,7 +43,7 @@ record contains a set of recorders for every specific label set used in the
 callback.
 
 A sync.Map maintains the mapping of current instruments and label sets to
-internal records.  To create a new bound instrument, the SDK consults the Map to
+internal records.  To find a record, the SDK consults the Map to
 locate an existing record, otherwise it constructs a new record.  The SDK
 maintains a count of the number of references to each record, ensuring
 that records are not reclaimed from the Map while they are still active